backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-generativeai
PyMuPDF
python-docx
python-pptx
chromadb
sentence-transformers
numpy
openai
requests
